Exercise04 - Simple SIFT Keypoint detection, description and matching. (#4)
Exercise04 - Simple SIFT Keypoint detection, description and matching

I guess there are still some bugs in my code, but because of
lack of time and relatively good results, I am going to the next exercise for now. I also skipped the
optional part of the exercise. I might come back to it later. The descriptor matching could be optimized later.